--- app/packages/core/src/components/Group/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Group";
