+++ app/packages/core/src/components/Modal/Group/DynamicGroup/NestedGroup/index.tsx
@@ -1,0 +1,7 @@
+import { useDynamicGroupContext } from "../DynamicGroupContextProvider";
+
+export const NestedGroup = () => {
+  const { lookerRefCallback } = useDynamicGroupContext();
+
+  return <div>dynamic</div>;
+};
